Refactor to implement review feedback
