+++ src/internet_identity/src/main.rs
@@ -1,6 +1,5 @@
 use crate::anchor_management::tentative_device_registration;
 use crate::assets::init_assets;
-use crate::state::ArchiveState;
 use candid::Principal;
 use ic_cdk::api::{caller, set_certified_data, trap};
 use ic_cdk_macros::{init, post_upgrade, pre_upgrade, query, update};
@@ -8,6 +7,7 @@ use ic_certified_map::AsHashTree;
 use serde_bytes::ByteBuf;
 use storage::{PersistentStateError, Salt, Storage};
 
+use crate::archive::ArchiveState;
 use internet_identity_interface::*;
 
 mod anchor_management;
